revert: rename context-plugin.ts from review
